Uploaded Piano Project
Create a piano that plays back assigned pitches. This project helped reinforce the concept of data.type in html and how they are linked into javascript. This also helped me gain a better understanding of the => arrow function and its benefits. In the css document, I also document how the use of css variables can be useful for reducing lines of code along with making responsive design easier and cleaner.

Updates needed. Needs more pitches. Timeout for active highlight needs to be shorter, explore other ways to write out that bit of code.

Upgrades coming. Added an instrument selector. Cycle through a preset list of sounds to play on the piano.